refac: WIP- tests pass, but Relationships not refacced
Tests test002 pass with refactored code (i.e., using convert.py:spec_to_entity()
- Nodes, Properties, Terms refactored; Relationships not yet

Union type deprecated and also removed from tests
MDF writer tests moved to new test suit test003; these tests are
marked to skip.
- to be done, move writer code into its own class MDFWriter.
