+++ python/tests/samples/test-model-union-type.yml
@@ -1,3 +1,4 @@
+# union type is deprecated 
 Nodes:
   case:
     Props:
